// frontend/src/app/core/api/generated/model/breakdownPerformanceIndicatorDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Statystyki związane z wydajnością w kontekście awarii
 */
export interface BreakdownPerformanceIndicatorDTO { 
    /**
     * Data, której dotyczą statystyki wydajności
     */
    date?: string;
    /**
     * Wartość wskaźnika wydajności (np. w procentach)
     */
    performance?: number;
}